refactor(agents): modularize base-sub-agent into domain-focused modules
SD-LEO-REFAC-BASE-AGENT-003: Refactor Base Sub-Agent

Extracted 850 LOC monolithic file into 5 focused modules:
- exceptions.js (54 LOC): Exception classes for budget/venture governance
- budget-manager.js (118 LOC): Supabase client, budget checking, logging
- finding-manager.js (175 LOC): Finding creation, deduplication, scoring
- output-generator.js (255 LOC): Output generation, status, recommendations
- index.js (297 LOC): Main BaseSubAgent class with factory pattern

Key changes:
- All modules under 500 LOC target (largest: 297 LOC)
- Backward-compatible wrapper preserves all 14 exports
- Added 12 comprehensive tests covering all modules
- Domain Extraction pattern ensures clean module boundaries

LEO Protocol: 4 handoffs accepted, retrospective created
